- rename classes - move classes threw packages
