+++ src/main/java/com/patryklikus/winter/beans/Beans.java
@@ -1,11 +1,11 @@
 /* Copyright patryklikus.com All Rights Reserved. */
 package com.patryklikus.winter.beans;
 
-import static java.lang.annotation.ElementType.TYPE;
-
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.TYPE;
 
 /**
  * All classes with this annotation will be searched to initialize objects using all its public methods. This class can inherit just from {@link Object}.
