+++ src/main/java/com/patryklikus/winter/lifecycle/Run.java
@@ -1,13 +1,13 @@
 /* Copyright patryklikus.com All Rights Reserved. */
 package com.patryklikus.winter.lifecycle;
 
-import static java.lang.annotation.ElementType.METHOD;
-import static java.util.concurrent.TimeUnit.SECONDS;
-
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import java.util.concurrent.TimeUnit;
+
+import static java.lang.annotation.ElementType.METHOD;
+import static java.util.concurrent.TimeUnit.SECONDS;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(METHOD)
