+++ src/test/java/com/patryklikus/winter/beans/exampleProject/config/Config.java
@@ -6,6 +6,7 @@ import com.patryklikus.winter.beans.exampleProject.models.Color;
 import com.patryklikus.winter.lifecycle.Close;
 import com.patryklikus.winter.lifecycle.Init;
 import com.patryklikus.winter.lifecycle.Run;
+
 import java.util.Arrays;
 import java.util.List;
